Moved the DeriveAddress files to the recovery package
