+++ pkg/extensions/tbtc/recovery/derive_address.go
@@ -1,4 +1,4 @@
-package tbtc
+package recovery
 
 import (
 	"fmt"
